RWA-1757 - Refactor classes to use consistent naming convention
